requirements.txt
flask
flask_sqlalchemy
Flask-Security-Too==5.1.2
flask_restful
celery
Flask-Excel
Jinja2
flask_caching
requests
redis